-added new result viewer dummy -x64 client setup is working again
git-svn-id: file:///root/svn_import/SolidSBC@62 bd7e3521-35e9-406e-9279-390287f868d3

+++ SolidSBCDatabaseLib/SolidSBCResultDBConnector.h
@@ -44,7 +44,7 @@ protected:
 public:
 	virtual int GetNameFromUuid(const CString& strUuid, CString& strName)=0;
 
-	int GetAllUniqueClients(std::map<CString,CString>& mapClientIdUuid) { mapClientIdUuid = m_mapClientIdUuid ; return m_mapClientIdUuid.size(); }
+	int GetAllUniqueClients(std::map<CString,CString>& mapClientIdUuid) { mapClientIdUuid = m_mapClientIdUuid ; return (int)m_mapClientIdUuid.size(); }
 
 protected:
 	int GetNameFromUuidSQLString(const CString& strUuid, CStringArray& arSQLCmds);
